local-app/requirements.txt
requests
pillow
psutil